Created an actual sort function instead of just initiating a new Deck. Cleaner this way.
The Sort function uses the id of the suit in the suits array along with the value of the card to determine the position of the card in the array. The function will work as long as all cards are unique. If there are multiple cards of the same suit and value, the two cards will overwrite eachother in the same spot in the array, as it stands now.